+++ internal/app/sla.go
@@ -168,8 +168,8 @@ func (s *SLAService) ListSLAs(ctx context.Context, req *cases.ListSLARequest) (*
 		Fields:  fields,
 		Context: ctx,
 		Sort:    req.Sort,
-		Page:    int64(page),
-		Size:    int64(req.Size),
+		Page:    int32(page),
+		Size:    int32(req.Size),
 		Time:    t,
 		Filter:  make(map[string]interface{}),
 	}
